+++ packages/ramps-controller/src/RampsService.test.ts
@@ -1961,6 +1961,7 @@ describe('RampsService', () => {
           quote: {
             amountIn: 100,
             amountOut: '0.05',
+            extraFee: '0',
             paymentMethod: '/payments/debit-credit-card',
             amountOutInFiat: 98,
           },
@@ -2035,6 +2036,7 @@ describe('RampsService', () => {
 
       expect(quotesResponse.success).toHaveLength(2);
       expect(quotesResponse.success[0]?.provider).toBe('/providers/moonpay');
+      expect(quotesResponse.success[0]?.quote.extraFee).toBe('0');
       expect(quotesResponse.sorted).toHaveLength(2);
       expect(quotesResponse.error).toHaveLength(0);
     });